Move BaseReferenceText to richtext module to reflect pybtex's module structure.
